requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.6.3
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
decorator==5.1.1
executing==0.8.3
frozenlist==1.3.0
googlesearch-python==1.1.0
idna==2.10
ipdb==0.13.9
ipython==8.4.0
jedi==0.18.1
joblib==1.1.0
matplotlib-inline==0.1.3
multidict==6.0.2
nltk==3.7
numpy==1.22.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
python-telegram-bot==13.12
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.6.2
requests==2.25.1
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.3.0
telepot==12.7
toml==0.10.2
torch==1.11.0
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
typing-extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
wcwidth==0.2.5
yarl==1.7.2
